// integration/src/api/sited_io/commerce/v2/offer_pb.ts
export class Offer extends Message<Offer> {
  /**
   * @generated from field: string offer_id = 1;
   */
  offerId = "";

  /**
   * @generated from field: string owner = 2;
   */
  owner = "";

  /**
   * @generated from field: int64 created_at = 3;
   */
  createdAt = protoInt64.zero;

  /**
   * @generated from field: int64 updated_at = 4;
   */
  updatedAt = protoInt64.zero;

  /**
   * @generated from field: sited_io.commerce.v2.Offer.Details details = 5;
   */
  details?: Offer_Details;

  /**
   * @generated from field: sited_io.commerce.v2.OfferType offer_type = 6;
   */
  offerType?: OfferType;

  /**
   * @generated from field: optional sited_io.commerce.v2.OfferPrice price = 7;
   */
  price?: OfferPrice;

  /**
   * @generated from field: repeated sited_io.commerce.v2.ShippingRate shipping_rates = 8;
   */
  shippingRates: ShippingRate[] = [];

  /**
   * @generated from field: repeated sited_io.commerce.v2.OfferImage images = 9;
   */
  images: OfferImage[] = [];

  constructor(data?: PartialMessage<Offer>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.Offer";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "offer_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "owner", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "created_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 4, name: "updated_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 5, name: "details", kind: "message", T: Offer_Details },
    { no: 6, name: "offer_type", kind: "message", T: OfferType },
    { no: 7, name: "price", kind: "message", T: OfferPrice, opt: true },
    { no: 8, name: "shipping_rates", kind: "message", T: ShippingRate, repeated: true },
    { no: 9, name: "images", kind: "message", T: OfferImage, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Offer {
    return new Offer().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Offer {
    return new Offer().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Offer {
    return new Offer().fromJsonString(jsonString, options);
  }

  static equals(a: Offer | PlainMessage<Offer> | undefined, b: Offer | PlainMessage<Offer> | undefined): boolean {
    return proto3.util.equals(Offer, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v2.Offer.Details
 */
export class Offer_Details extends Message<Offer_Details> {
  /**
   * @generated from field: string name = 1;
   */
  name = "";

  /**
   * @generated from field: optional string description = 2;
   */
  description?: string;

  constructor(data?: PartialMessage<Offer_Details>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.Offer.Details";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "description", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Offer_Details {
    return new Offer_Details().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Offer_Details {
    return new Offer_Details().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Offer_Details {
    return new Offer_Details().fromJsonString(jsonString, options);
  }

  static equals(a: Offer_Details | PlainMessage<Offer_Details> | undefined, b: Offer_Details | PlainMessage<Offer_Details> | undefined): boolean {
    return proto3.util.equals(Offer_Details, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v2.OfferType
 */
export class OfferType extends Message<OfferType> {
  /**
   * @generated from oneof sited_io.commerce.v2.OfferType.offer_type_kind
   */
  offerTypeKind: {
    /**
     * @generated from field: sited_io.commerce.v2.OfferType.Physical physical = 1;
     */
    value: OfferType_Physical;
    case: "physical";
  } | {
    /**
     * @generated from field: sited_io.commerce.v2.OfferType.Digital digital = 2;
     */
    value: OfferType_Digital;
    case: "digital";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<OfferType>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.OfferType";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "physical", kind: "message", T: OfferType_Physical, oneof: "offer_type_kind" },
    { no: 2, name: "digital", kind: "message", T: OfferType_Digital, oneof: "offer_type_kind" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferType {
    return new OfferType().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferType {
    return new OfferType().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferType {
    return new OfferType().fromJsonString(jsonString, options);
  }

  static equals(a: OfferType | PlainMessage<OfferType> | undefined, b: OfferType | PlainMessage<OfferType> | undefined): boolean {
    return proto3.util.equals(OfferType, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v2.OfferType.Physical
 */
export class OfferType_Physical extends Message<OfferType_Physical> {
  constructor(data?: PartialMessage<OfferType_Physical>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.OfferType.Physical";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferType_Physical {
    return new OfferType_Physical().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferType_Physical {
    return new OfferType_Physical().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferType_Physical {
    return new OfferType_Physical().fromJsonString(jsonString, options);
  }

  static equals(a: OfferType_Physical | PlainMessage<OfferType_Physical> | undefined, b: OfferType_Physical | PlainMessage<OfferType_Physical> | undefined): boolean {
    return proto3.util.equals(OfferType_Physical, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v2.OfferType.Digital
 */
export class OfferType_Digital extends Message<OfferType_Digital> {
  constructor(data?: PartialMessage<OfferType_Digital>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.OfferType.Digital";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferType_Digital {
    return new OfferType_Digital().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferType_Digital {
    return new OfferType_Digital().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferType_Digital {
    return new OfferType_Digital().fromJsonString(jsonString, options);
  }

  static equals(a: OfferType_Digital | PlainMessage<OfferType_Digital> | undefined, b: OfferType_Digital | PlainMessage<OfferType_Digital> | undefined): boolean {
    return proto3.util.equals(OfferType_Digital, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v2.OfferPrice
 */
export class OfferPrice extends Message<OfferPrice> {
  /**
   * @generated from field: sited_io.price.v1.Price price = 1;
   */
  price?: Price;

  /**
   * @generated from field: sited_io.commerce.v2.PriceType price_type = 2;
   */
  priceType?: PriceType;

  constructor(data?: PartialMessage<OfferPrice>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.OfferPrice";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "price", kind: "message", T: Price },
    { no: 2, name: "price_type", kind: "message", T: PriceType },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferPrice {
    return new OfferPrice().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferPrice {
    return new OfferPrice().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferPrice {
    return new OfferPrice().fromJsonString(jsonString, options);
  }

  static equals(a: OfferPrice | PlainMessage<OfferPrice> | undefined, b: OfferPrice | PlainMessage<OfferPrice> | undefined): boolean {
    return proto3.util.equals(OfferPrice, a, b);
  }
}
// ...
export class ShippingRate extends Message<ShippingRate> {
  /**
   * @generated from field: string shipping_rate_id = 1;
   */
  shippingRateId = "";

  /**
   * @generated from field: string owner = 2;
   */
  owner = "";

  /**
   * @generated from field: sited_io.price.v1.Price price = 3;
   */
  price?: Price;

  /**
   * @generated from field: bool all_countries = 4;
   */
  allCountries = false;

  /**
   * @generated from field: repeated sited_io.country.v1.CountryCode specific_countries = 5;
   */
  specificCountries: CountryCode[] = [];

  constructor(data?: PartialMessage<ShippingRate>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v2.ShippingRate";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "shipping_rate_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "owner", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "price", kind: "message", T: Price },
    { no: 4, name: "all_countries", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 5, name: "specific_countries", kind: "enum", T: proto3.getEnumType(CountryCode), repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ShippingRate {
    return new ShippingRate().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ShippingRate {
    return new ShippingRate().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ShippingRate {
    return new ShippingRate().fromJsonString(jsonString, options);
  }

  static equals(a: ShippingRate | PlainMessage<ShippingRate> | undefined, b: ShippingRate | PlainMessage<ShippingRate> | undefined): boolean {
    return proto3.util.equals(ShippingRate, a, b);
  }
}