// integration/src/api/sited_io/commerce/v2/query_pb.ts
export enum OffersOrderByField {
  /**
   * @generated from enum value: OFFERS_ORDER_BY_FIELD_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: OFFERS_ORDER_BY_FIELD_CREATED_AT = 1;
   */
  CREATED_AT = 1,

  /**
   * @generated from enum value: OFFERS_ORDER_BY_FIELD_UPDATED_AT = 2;
   */
  UPDATED_AT = 2,
}
// ...
export enum OffersFilterField {
  /**
   * @generated from enum value: OFFERS_FILTER_FIELD_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: OFFERS_FILTER_FIELD_NAME = 1;
   */
  NAME = 1,

  /**
   * @generated from enum value: OFFERS_FILTER_FIELD_DESCRIPTION = 2;
   */
  DESCRIPTION = 2,

  /**
   * @generated from enum value: OFFERS_FILTER_FIELD_NAME_AND_DESCRIPTION = 3;
   */
  NAME_AND_DESCRIPTION = 3,

  /**
   * @generated from enum value: OFFERS_FILTER_FIELD_TYPE = 4;
   */
  TYPE = 4,
}